Separate task resolution from provider setup, and refuse mistyped task paths
`session.rs` held three unrelated things: turning `--task` into text, finding
and running an editor, and building the provider registry. The editor half just
moved to `leviath-sys`; this moves the task half into `run/task.rs` and leaves
`session.rs` doing one job.

`read_region_value` comes along, since resolving `--<region> @file` is the same
concern as resolving `--task <file>` and the two are documented against each
other.

Three signatures get simpler on the way. `arg` becomes `Option<&str>`, which is
what the binary already has and drops a clone. `description` becomes `&str`,
because the only production source is `Blueprint::description`, which the
manifest parser defaults to "" - the `None` arm was unreachable. And the editor
seam now speaks `io::Result`, so `leviath_sys::editor::launch` coerces straight
in with no wrapper.

The behavior change: `--task ./promt.md` used to spawn an agent whose entire
task was the string "./promt.md". Real tokens, a run id in `lev ps`, and no
signal that anything was wrong. A value that has no whitespace and does carry a
separator is now required to name something. Prose that mentions a path ("fix
src/main.rs") has spaces, so it is untouched, and a bare word with no separator
is still literal.
